+++ components/Button/SignButton/SignButton.module.scss
@@ -1,0 +1,3 @@
+.sign-in {}
+
+.sign-out {}
